# docs/06-API/openapi.yaml
openapi: 3.1.0
info:
  title: ProjectPulse MCP API
  version: 2.0.0
  description: |
    **ProjectPulse MCP API Specification**

    Complete API specification for the ProjectPulse Model Context Protocol (MCP) server. This API exposes 45 tools across 9 functional categories, enabling AI agents to manage software development workflows with 95% automation.

    **Architecture:** Agent-First (95% MCP automation, 5% UI monitoring)
    **Primary Users:** AI Agents (Claude Code, Cursor AI, Codex)
    **Secondary Users:** Solo/small team developers (monitoring, overrides)

    **Key Features:**
    - 45 MCP tools organized into 9 categories
    - Complete CRUD operations for all entities
    - Token-optimized responses (92% reduction for skills, 88% for knowledge)
    - Autonomy levels (0-4) for security control
    - Full audit trail via AgentAction logging

    **Standards:**
    - OpenAPI 3.1.0
    - JSON Schema for validation
    - RESTful principles
    - MCP stdio transport

    **Related Documentation:**
    - [AgentOps Plan](../05-AgentOps-Plan.md) - Complete tool catalog with workflows
    - [SRS](../02-SRS.md) - Functional requirements (FR-001 to FR-125)
    - [Architecture](../03-Architecture.md) - System design and data flows
    - [Data Model](../04-Data-and-Model-Spec.md) - Database schema

  contact:
    name: ProjectPulse Team
    email: team@projectpulse.dev
    url: https://projectpulse.dev

  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
...
security:
  - autonomyLevel: []
...
paths:
...
  /onboarding/blueprint:
    get:
      operationId: onboarding.blueprint.get
      summary: Get project blueprint (Sprint 8.5)
      description: |
        Retrieve the Session 3 blueprint data (memory banks, skills, SOPs) for agent initialization.
        Used by `projectpulse.blueprint.get` MCP tool.

        **Requirements:** FR-028
      tags:
        - Sprint Tracking
      security:
        - autonomyLevel: [0]
      responses:
        '200':
          description: Blueprint data retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  memoryBanks:
                    type: array
                    items:
                      type: object
                  skills:
                    type: array
                    items:
                      type: object
                  sops:
                    type: array
                    items:
                      type: object
...
  /workflow/complete:
    post:
      operationId: workflow.complete
      summary: Complete workflow (Step 5)
      description: |
        **5-Step Protocol - Step 5: Post-Completion**

        Mark workflow complete in database; update progress and session state.
        Optionally create completion summary record.
        Changes immediately visible in Development Cycle UI and via MCP tools.
        Optionally invoke synthesize-docs and map-system sub-agents.

        **Requirements:** FR-034, FR-035
      tags:
        - Workflow Orchestration
      security:
        - autonomyLevel: [1]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - workflowId
                - completionSummary
              properties:
                workflowId:
                  type: integer
                  example: 456
                completionSummary:
                  type: string
                  example: 'Issue management API complete with 8 endpoints'
                invokeDocAgents:
                  type: boolean
                  description: Invoke synthesize-docs + map-system if true
                  example: true
      responses:
        '200':
          description: Workflow completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  workflowComplete:
                    type: boolean
                    example: true
                  completionSummaryId:
                    type: integer
                    description: Optional DB record for completion summary
                    example: 123
                  progressUpdated:
                    type: boolean
                    description: Task/Session progress updated in DB
                    example: true
                  step5Complete:
                    type: boolean
                    example: true
...
components:
  securitySchemes:
    autonomyLevel:
      type: apiKey
      in: header
      name: X-Autonomy-Level
      description: |
        Autonomy level for agent operations (0-4):
        - **Level 0 (Read-Only):** Read data, list resources
        - **Level 1 (Safe Writes):** Create issues, update status
        - **Level 2 (Approval Required):** Delete issues, modify schema
        - **Level 3 (Infrastructure):** Deploy, modify production
        - **Level 4 (Full Autonomy):** All operations (future)
...